Split editor and UI class reusing shared methods.
